+++ src/components/Footer.tsx
@@ -1,7 +1,7 @@
 export default function FooterGlobal() {
 	return (
 		<footer>
-			<div class="bg-laidlawGrey w-full text-sm flex justify-center items-center p-6 gap-6 text-white relative">
+			<div className="bg-laidlawGrey w-full text-sm flex justify-center items-center p-6 gap-6 text-white relative">
 				Copyright © 2024 Laidlaw Opera Trust. All rights reserved.
 			</div>
 		</footer>
