+++ src/layouts/DefaultPageLayout.astro
@@ -1,8 +1,8 @@
 ---
 // define functions between ---
 import "@fontsource/josefin-sans/400.css";
-import HeaderGlobal from "../components/Header";
-import FooterGlobal from "../components/Footer";
+import HeaderGlobal from "../components/Header.tsx";
+import FooterGlobal from "../components/Footer.tsx";
 
 interface Props {
 	title: string;
